refactor: 댓글의 수정, 삭제는 댓글의 작성자 혹은 게시글의 작성자만 가능합니다.

+++ src/main/java/com/sparta/newsfeed/comment/service/CommentService.java
@@ -12,6 +12,7 @@ import com.sparta.newsfeed.post.entity.Post;
 import com.sparta.newsfeed.post.repository.PostRepository;
 import com.sparta.newsfeed.user.entity.User;
 import com.sparta.newsfeed.user.repository.UserRepository;
+import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -50,19 +51,38 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(Long commentId, CommentRequestDto commentRequestDto) {
+    public void updateComment(HttpServletRequest request, Long commentId, CommentRequestDto commentRequestDto) {
+        // 댓글은 작성자 혹은 게시물의 작성자만 수정이 가능합니다.
+        User user = userRepository.findById(getUserId(request)).orElseThrow(() -> new ApplicationException(ErrorCode.USER_NOT_FOUND));
+
         Comment comment = findComment(commentId);
 
-        comment.update(commentRequestDto.getContent());
-
+        // 댓글 작성자의 Id와 사용자의 Id, 게시물 작성자의 Id와 비교
+        if((user.getId() == comment.getUser().getId()) || (user.getId() == comment.getPost().getId())) {
+            // 수정을 요청한 유저Id가 댓글의 유저Id 또는 게시물의 유저Id와 일치한다면 -> 수정 가능
+            comment.update(commentRequestDto.getContent());
+        }else{
+            // Id가 일치하지 않으면 -> 수정 불가
+            throw new ApplicationException(ErrorCode.USER_CANNOTUPDATE_COMMENT);
+        }
         commentRepository.save(comment);
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(HttpServletRequest request, Long commentId) {
+        // 댓글은 작성자 혹은 게시물의 작성자만 수정이 가능합니다.
+        User user = userRepository.findById(getUserId(request)).orElseThrow(() -> new ApplicationException(ErrorCode.USER_NOT_FOUND));
+
         Comment comment = findComment(commentId);
 
-        commentRepository.delete(comment);
+        // 댓글 작성자의 Id와 사용자의 Id, 게시물 작성자의 Id와 비교
+        if((user.getId() == comment.getUser().getId()) || (user.getId() == comment.getPost().getId())) {
+            // 삭제을 요청한 유저Id가 댓글의 유저Id 또는 게시물의 유저Id와 일치한다면 -> 삭제 가능
+            commentRepository.delete(comment);
+        }else{
+            // Id가 일치하지 않으면 -> 수정 불가
+            throw new ApplicationException(ErrorCode.USER_CANNOTDELETE_COMMENT);
+        }
     }
 
     public Comment findComment(Long commentId) {
